requirements.txt
pillow
requests
python-slugify